Added Main HomeScreen and Linking
